react-full-post: Moved top CheckMyPost and Nav components to new file
