+++ public/js/post.js
@@ -1,11 +1,11 @@
 requirejs([
-    'es6!post/post_components'],
+    'es6!post/post_page'],
     function (
-        postComponents
+        checkMyPost
     ) {
 
     $(document).ready(function () {
-        ReactDOM.render(React.createElement(postComponents.CheckMyPost, null), document.getElementById('all_posts'));
+        ReactDOM.render(React.createElement(checkMyPost.CheckMyPost, null), document.getElementById('all_posts'));
         // openLastActiveTab();
     });
 });
